% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{sft_getIslands}
\alias{sft_getIslands}
\title{Get small detached polygons (islands)}
\usage{
sft_getIslands(x, forget = 0)
}
\arguments{
\item{x}{sf or sfc object}

\item{forget}{Number of areas not considered as islands}
}
\value{
same type than x. sf or sfc object
}
\description{
Get small detached polygons (islands)
}
\examples{
library(sftoolkit)
library(ggplot2)
data("Tasmania")

Islands <- sft_getIslands(Tasmania)
ggplot() + geom_sf(data=Tasmania, fill=NA, color="black", size = 0.25)+theme_light() +
 geom_sf(data=Islands, fill=NA, color="red", size = 0.25)+theme_light()
}
