% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sft_innerLines.R
\name{sft_innerLines}
\alias{sft_innerLines}
\title{Create a line layer consisting of shared boundaries with no attribute data}
\usage{
sft_innerLines(x)
}
\arguments{
\item{x}{sf or sfc polygons object}
}
\value{
multilines in the same class as the input layer, but without attributes
}
\description{
Create a line layer consisting of shared boundaries with no attribute data
}
\examples{
library(sftoolkit)
data("Tasmania")
innerlines <- sftk_innerlines(Tasmania)
plot(innerlines)
}
\author{
Jean Pierre Decorps <jp.decorps@epiconcept.fr>
}
