% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{sft_removeIslands}
\alias{sft_removeIslands}
\title{Remove small detached polygons (islands)}
\usage{
sft_removeIslands(x, min_area = 4)
}
\arguments{
\item{x}{spatial object to filter of sf or sfc}

\item{min_area}{minimum area to keep (square kilometer)}
}
\value{
like x, sf or sfc
}
\description{
Remove small detached polygons, keeping those with a minimum area
}
