% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tasmania}
\alias{Tasmania}
\title{Map of Tasmania (AUS)}
\format{sf, data frame with 30 rows and 6 variables + geometry.
\describe{
  \item{ISO}{ISO code of Australia: AUS}
  \item{NAME_0}{"Australia"}
  \item{NAME_1}{"Tasmania"}
  \item{NAME_2}{Names of Municipalities}
  \item{TYPE_2}{"Municipality"}
  \item{ENGTYPE_2}{"Municipality"}
  \item{geometry}{sfc_MULTIPOLYGON}
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
Tasmania
}
\description{
An object of class "sf" "data.frame" with geometry and other attributes
of the map of Tasmnia.
}
\keyword{datasets}
