% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{sft_removeIslands}
\alias{sft_removeIslands}
\title{Remove small detached polygons (islands)}
\usage{
sft_removeIslands(x, mode = "keeplarger", min_area = 4, keep = 1)
}
\arguments{
\item{x}{spatial object to filter of sf or sfc}

\item{mode}{can be one of "keeplarger" "bysize", "byvertices"}

\item{min_area}{minimum area to keep (square kilometer)}

\item{keep}{number of polygons to keep when mode = "keeplarger" and no sub-regions}
}
\value{
like x, sf or sfc
}
\description{
Remove small detached polygons, keeping those with a minimum area
}
